Reduce warnings in REPL
There is some overlap between -Xlint and -Ywarn-unused. One of the
things -Xlint enables is -Xlint:unused, which in turn enables some of
what -Ywarn-unused does, but not all. When starting the REPL I
disabled -Ywarn-unused, but because -Xlint:unused is in the mix, we
still got warnings about unused imports.

Because -Ywarn-unused is a superset of -Xlint:unused, I just disabled
the latter. No more warnings in the REPL.

See 'scalac -Xlint:help' and 'scalac -Ywarn-unused:help' for more.