Add request, response types for info requests
We will need to revisit where these types live after this is merged
and the project restructuring is done.

There are also some things that should be done to make these types
better capture the requirements in the spec. For instance, `Bin` types
cannot have both the `ranges` and `centers` fields defined. It would
be good to enforce this somehow.

See #8.