Make HAPI service more suitable for standalone use
This commit moves some responsibilities from the HAPI server
implementation to the HAPI service so the service can be more easily
pulled into other projects.

Prior to this commit, the HAPI server implementation was responsible
for configuring and combining the individual services implementing the
HAPI endpoints. Because we want a reusable HAPI service, I've
repurposed the `HapiService` class (and moved the existing landing
page stuff) to make it responsible for the configuration and
coordination of the individual services that implement the HAPI
endpoints. This way, a server that wishes to include a HAPI service
just needs to get the service from the `HapiService` class and mount
it.

The spec requires that the endpoints appear directly after a `hapi`
path element in the URL, so I've changed the routes for the services
implementing each endpoint to enforce this.