Construct DataRequest object from query parameters
This commit sets up the path from query parameters to a `DataRequest`
object, which will eventually be used to fulfill the request.

Part of this path involves validating that the start and stop times
have the appropriate format and checking that the min time is an
instant before the max time (#26). We also check that the `include`
and `format` parameters have reasonable arguments if they are
specified (#28). If any of those conditions are not met, we'll return
an appropriate HAPI error.