Add a catalog algebra and interpreter
This commit adds an algebra for reading catalogs with one operation to
do the reading. I assume catalogs will be small, so the operation
returns the catalog as a list. The HAPI spec doesn't define any query
for the catalog endpoint, so just getting the catalog is all that
needs to be done.

I've also implemented an interpreter for this algebra that just
returns an empty catalog. It's a useful default because it makes
things compile but doesn't actually do anything.

I imagine the algebras for each service being implemented in a single
interpreter that is passed to `HapiService`, just like how all the
individual http4s services making up a whole HAPI service are bundled.

See #7.