Use NonEmptyList for dataset parameter subsets
Queries with empty parameter lists (i.e. `parameters=`) are invalid
and we return 400 errors for those requests. It then follows that the
internal representation of the requested subset of parameters is a
non-empty list. Either no subset was specified, or we have a non-empty
list of the parameters in the subset.

The response cannot contain an empty parameter list because the first
parameter must always be a time value. Using a non-empty list ensures
that we at least have one parameter in the response, though it would
be better to ensure that we have at least one parameter and it is the
time parameter.