Add an info algebra
The info algebra is more complicated than the catalog algebra because
the info service takes request parameters. The only operation is
getting metadata for a particular dataset (by ID) and optionally for
only a subset of its parameters.

There are two error cases when asking for metadata which are
represented by the `InfoError` sum type. The first is asking for
metadata for a dataset ID that doesn't exist. The second is asking for
a parameter that doesn't exist. The corresponding types capture the ID
or parameter name that triggered the error because I intend to log
these internally. They aren't (and shouldn't be) used to generate
user-facing errors.

Because we may produce a valid result or a particular error in some F
context, I've used `EitherT` as the result type of this operation.

See #10.