Separate dataset metadata from InfoResponse type
The `InfoResponse` type is intended to represent the service's
response to an `info` request. Logically, the response is the version
of the HAPI server, the status code, and the dataset metadata. This
commit changes the `InfoResponse` type to better reflect that.

By having a separate `Metadata` type, interpreters of `InfoAlgebra`
need only be concerned about producing an instance of `Metadata`
rather than the entirety of `InfoResponse`. This is similar to how
interpreters of `CatalogAlgebra` need only produce `List[Dataset]`
rather than `Catalog` (though the naming here is poor, we should
consider renaming that to `CatalogResponse`).

The `format` field is an unfortunate case. The spec includes it in the
`info` response when it's only required when this information is part
of a `data` response. Because this is `InfoResponse`, I've removed it.
Perhaps a better way of thinking about a `data` response is that it
optionally has the format of the data and the metadata.