Specify version in error responses
Error responses from the HAPI server ought to specify the version of
the spec being used. This commit adds a `HapiError` wrapper over the
`Status` type that includes the version when encoded as JSON.

I introduced `HapiError` rather than change the encoder for `Status`
because most of the time we use `Status` we don't want to include the
version because it's already represented elsewhere. It's only the
error cases that didn't have the version associated with them.

Fixes #38.