Fix reading catalogs from LaTiS 2
I thought this worked, but running against the LISIRD catalog (and
against my original test catalog) fails without these changes.

I don't understand nested catalogs in LaTiS 2, so this is something
we'll need to revisit. However, we don't actually use nested catalogs
anywhere, so it's unlikely to be a priority in the near future.