Implement info algebra in interpreters
This commit does the following:

- Implements `InfoService` in terms of `InfoAlgebra`.

- Makes `HapiInterpreter` extend `InfoAlgebra`.

- Implements `InfoAlgebra` in the default interpreter by always
  returning an error for an unknown ID. This is the only thing that
  makes sense (for now) because the catalog is empty, so all dataset
  IDs are unknown.

- Implements `InfoAlgebra` in the LaTiS 2 interpreter.

We can now access metadata from LaTiS 2 datasets, subject to the
following limitations:

- The dataset is a simple time series (i.e. no nested functions).

- The domain is a scalar time.

- The time variable has the name "time".

- The time variable is text in the HAPI time format. This is so the
  length metadata is correct. Other time formats will work, including
  text times of different formats, but the length may be incorrect or
  not present. (Length is always required for times.)

- The range is a scalar or a tuple containing only scalars.

- There is an unfortunate special case for projecting only time. When
  projecting a single variable `X`, LaTiS returns `index -> X`. So,
  when projecting time alone, time shows up in the range rather than
  the domain. This works, but the code is more complicated.

Some additional commentary: By making `HapiInterpreter` extend
`InfoAlgebra`, we force all interpreters extending `HapiInterpreter`
to implement `InfoAlgebra`, regardless of whether we need to use it
for accessing metadata yet. We also can't implement the service in
terms of its algebra because the implementation is required as a
constructor parameter. This means we need to update everything at one
in order to have an atomic commit that leaves the repository in a
state that compiles, which probably isn't desirable.

See #10.