Add data algebra
I've decided to represent the type of the data being requested using a
type member rather than a type parameter to avoid needing to specify
that detail everywhere.

I started with `DataAlgebra[F[_], T]`, but that required going all the
way up to `HapiService` and redefining it to include `T`. That struck
me as wrong, because at the service level we don't care about the type
of data we're getting from the backend.

See #11.