Implement data algebra in interpreters
This commit does the following:

- Implements the data reading and writing parts of `DataService` in
  terms of `DataAlgebra`.

- Makes `HapiInterpreter` extend `DataAlgebra`.

- Implements `DataAlgebra` in the default interpreter. Asking for data
  yields unit and trying to write data yields an empty stream.

- Implements `DataAlgebra` in the LaTiS 2 interpreter.

We can now read data through LaTiS 2 and write it out as CSV subject
to the same constraints as are on the info response. Briefly, this
will only work if the dataset is a simple time series (t -> a or t
-> (a, b, ...) without nested tuples). However, times will be
converted to HAPI time strings.

See #11.