Set up configuration
This commit introduces a type for keeping our configuration, makes the
server read the configuration when starting, and introduces a default
configuration.

The idiom appears to be to have a `reference.conf` file on the
classpath that contains defaults. This configuration provides defaults
but also specifies environment variables that could be used to
override these. The defaults can also be overridden with an
`application.conf` file provided at runtime.

We can't pass our configuration object to code using LaTiS 2 because
LaTiS 2 gets everything from `LatisProperties`. The easiest thing I
thought to do was to take the configuration values we get and set the
relevant system properties before using LaTiS 2.

See #14.