Handle edge cases in CSV decoding
This decoder should now reject following:

- empty inputs
- inputs consisting only of commas
- inputs with leading or trailing commas
- inputs with empty fields (adjacent commas)

The call to `.get` in the original was not safe because there was a
condition under which `split` would return an empty list: when the
input consists only of commas.

By adding `-1` as a second argument, no fields are ignored and `split`
will always return a non-empty list, even for empty inputs.

Fixes #40.