Clean up JSON responses
This commit cleans up our JSON responses by removing unset optional
parameters from the response.

Many parts of the info response are optional, and the default encoders
translate None into null, so we had large JSON objects full of nulls.
The new encoders introduced by this commit remove keys with null
values, with the exception of "fill" which is required to be present
even when null.

I've also had to explicitly rename "dType" to "type" because the
derived encoder will use the name of the parameter from the case
class.

See #45.