Make LaTiS 2 the default interpreter
Most of the changes required to make LaTiS 2 the default were finding
the appropriate classpath to copy into the Docker image.

The confusing part was the fact that we're depending on LaTiS 2 as a
source dependency. Apparently we can get the JAR through
`internalDependencyAsJars`, so I just add that to the things being
copied into the image.

See #43.