Support datasets with a second dimension
These changes enable the server to serve datasets with an additional
dimension. These will likely be things like spectral time series.

An unfortunate consequence of how we project variables and how HAPI
represents the inner domain is that we need to be careful to always
keep that inner domain when trying to get the metadata. I had to
change the way we get metadata from projecting only the requested
parameters to keeping all of them and then filtering out the ones we
didn't ask for so we always had the inner domain for creating bins.

See #52.