Clean up JSON for bins
Bins have optional parameters that should be removed from the JSON if
not specified.

However, the `centers` parameter is special in that a null is expected
if `ranges` is not set (but removed from the JSON) to indicate that
the dimension corresponding to the bin does not represent binned data.