Split server and service into separate projects
The major difficulty with this split was making sure that
sbt-revolver, sbt-docker, and sbt-assembly still did what we wanted
them to.

I needed to disable sbt-revolver on the root project and the service
project because they have nothing to run.

I needed to enable sbt-docker on the server project, because that's
the one we want to build the Docker image for.

Dealing with sbt-assembly was a little trickier. By default it will
build a JAR for each project. By making the root project depend on the
others and disabling aggregation for the assembly task, I was able to
convince it to produce a single JAR that pulls in everything from the
server and service projects.

I have no idea if this is the best way of doing this, but it appears
to work for our needs at the moment.

See #57.