Properly handle LaTiS properties when making executable JAR
Adding a `latis.properties` file to this project broke the `assembly`
task because it didn't know how to handle the one in this project and
the one included in the LaTiS 2 dependency.

I've changed the merge strategy to take the first `latis.properties`
in the classpath, which ought to be ours.