Take interpreter as argument when constructing app
This allows those with custom interpreters to run a server using their
interpreter without needing to duplicate some of the existing server
code.

I prefer this to loading classes specified in the configuration
because I think it's unlikely that typical users will want to swap out
the interpreter. A power user writing their own interpreter will
probably be fine writing a few extra lines of code to make the server
use it.

Closes #64.