Improve logging
The logging middleware that comes with http4s does most of what I
think we need. We can log the request and the response (without the
body) and use those to know what is being asked for and estimate how
long it took to generate the response.

Because we sometimes return large amounts of data that we don't want
to put in our logs, I've disabled logging the bodies of requests and
responses. This doesn't matter for requests, but for responses it
means we only get to see the HTTP error code and not the HAPI error
code.

To get the HAPI error codes I've added loggers to the info and data
services (the only ones that produce HAPI errors) that log the
associated HAPI error messages.

See #60.