Migrate to LaTiS v3
We are dropping support for LaTiS v2 and focusing on LaTiS v3. These
changes are to migrate the HAPI service interface to fit in the LaTiS
v3 service interface architecture, add an implementation of the HAPI
algebra for LaTiS v3, and make any tweaks to the HAPI service required
for LaTiS v3 support.

In order to support producing a standalone executable, the SBT build
has been modified to include dependencies on everything required to
make a standalone server. This is a temporary hack to work around the
fact that we don't have artifacts published that LaTiS can pull in.